requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
autogen-agentchat==0.7.2
autogen-core==0.7.2
banks==2.2.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
dashscope==1.24.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
faiss-cpu==1.11.0.post1
fastapi==0.127.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.4
griffe==1.11.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
langchain==1.2.3
langchain-community==0.3.27
langchain-core==1.2.7
langchain-openai==0.3.29
langchain-text-splitters==0.3.9
langgraph==1.0.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.4.13
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.13.1
llama-index-cli==0.5.0
llama-index-core==0.13.1
llama-index-embeddings-dashscope==0.4.0
llama-index-embeddings-openai==0.5.0
llama-index-indices-managed-llama-cloud==0.9.1
llama-index-instrumentation==0.4.0
llama-index-llms-openai==0.5.2
llama-index-llms-openai-like==0.5.0
llama-index-readers-file==0.5.0
llama-index-readers-llama-parse==0.5.0
llama-index-workflows==1.3.0
llama-parse==0.6.54
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.3.2
openai==1.99.1
opentelemetry-api==1.36.0
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
pandas==2.2.3
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
protobuf==5.29.5
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pypdf==5.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==7.1.0
regex==2025.7.33
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.6.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
starlette==0.50.0
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.21.4
tqdm==4.67.1
transformers==4.55.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uuid_utils==0.13.0
uv==0.9.18
uvicorn==0.40.0
websocket-client==1.8.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
